@@ -7,7 +7,7 @@ name = "nbd-server"
 version = "0.1.0"
 description = "Network Block Device (NBD) server with S3 backend and Copy-on-Write support"
 readme = "README.md"
-requires-python = ">=3.11"
+requires-python = ">=3.13"
 license = {text = "MIT"}
 authors = [
     {name = "Teja Chava"},
@@ -33,7 +33,7 @@ dev = [
 
 [tool.black]
 line-length = 100
-target-version = ['py311']
+target-version = ['py313']
 include = '\.pyi?$'
 
 [tool.isort]
@@ -42,7 +42,7 @@ line_length = 100
 multi_line_output = 3
 
 [tool.mypy]
-python_version = "3.11"
+python_version = "3.13"
 warn_return_any = true
 warn_unused_configs = true
 disallow_untyped_defs = true
@@ -82,7 +82,7 @@ exclude_lines = [
 
 [tool.ruff]
 line-length = 100
-target-version = "py311"
+target-version = "py313"
 select = [
     "E",   # pycodestyle errors
     "W",   # pycodestyle warnings
