@@ -1,5 +1,5 @@
 from .server import NBDServer
-from .storage import StorageBackend, InMemoryStorage
+from .storage import InMemoryStorage, StorageBackend
 
 __version__ = "0.1.0"
 __all__ = ["NBDServer", "StorageBackend", "InMemoryStorage"]
