@@ -1,7 +1,7 @@
 # NBD Protocol Magic Numbers
-NBDMAGIC = 0x4e42444d41474943  # "NBDMAGIC" in ASCII
-IHAVEOPT = 0x49484156454f5054  # "IHAVEOPT" in ASCII
-NBD_REP_MAGIC = 0x3e889045565a9
+NBDMAGIC = 0x4E42444D41474943  # "NBDMAGIC" in ASCII
+IHAVEOPT = 0x49484156454F5054  # "IHAVEOPT" in ASCII
+NBD_REP_MAGIC = 0x0003E889045565A9
 NBD_SIMPLE_REPLY_MAGIC = 0x67446698
 NBD_REQUEST_MAGIC = 0x25609513
 
@@ -31,6 +31,6 @@ NBD_CMD_DISC = 2  # Disconnect
 NBD_CMD_FLUSH = 3
 
 # Server Configuration
-DEFAULT_HOST = 'localhost'
+DEFAULT_HOST = "localhost"
 DEFAULT_PORT = 10809
 DEFAULT_EXPORT_SIZE = 1024 * 1024 * 1024  # 1GB
