@@ -3,24 +3,24 @@ import struct
 import threading
 import unittest
 
-from nbd_server.server import NBDServer
-from nbd_server.storage import InMemoryStorage
 from nbd_server.constants import (
-    NBDMAGIC,
     IHAVEOPT,
+    NBD_CMD_DISC,
+    NBD_CMD_FLUSH,
+    NBD_CMD_READ,
+    NBD_CMD_WRITE,
     NBD_FLAG_FIXED_NEWSTYLE,
     NBD_OPT_ABORT,
     NBD_OPT_GO,
-    NBD_REP_MAGIC,
-    NBD_REP_INFO,
     NBD_REP_ACK,
-    NBD_CMD_READ,
-    NBD_CMD_WRITE,
-    NBD_CMD_FLUSH,
-    NBD_CMD_DISC,
-    NBD_SIMPLE_REPLY_MAGIC,
+    NBD_REP_INFO,
+    NBD_REP_MAGIC,
     NBD_REQUEST_MAGIC,
+    NBD_SIMPLE_REPLY_MAGIC,
+    NBDMAGIC,
 )
+from nbd_server.server import NBDServer
+from nbd_server.storage import InMemoryStorage
 
 NBD_FLAG_C_FIXED_NEWSTYLE = 0x00000001
 
@@ -48,7 +48,7 @@ class TestNBDServerHandshake(unittest.TestCase):
             handshake_data = client_socket.recv(18)
             self.assertEqual(len(handshake_data), 18)
 
-            magic, ihaveopt, flags = struct.unpack('>QQH', handshake_data)
+            magic, ihaveopt, flags = struct.unpack(">QQH", handshake_data)
             self.assertEqual(magic, NBDMAGIC)
             self.assertEqual(ihaveopt, IHAVEOPT)
             self.assertEqual(flags, NBD_FLAG_FIXED_NEWSTYLE)
@@ -79,9 +79,9 @@ class TestNBDServerClientFlags(unittest.TestCase):
 
             client_socket.recv(18)
 
-            client_socket.sendall(struct.pack('>I', NBD_FLAG_C_FIXED_NEWSTYLE))
+            client_socket.sendall(struct.pack(">I", NBD_FLAG_C_FIXED_NEWSTYLE))
 
-            option_header = struct.pack('>QII', IHAVEOPT, NBD_OPT_ABORT, 0)
+            option_header = struct.pack(">QII", IHAVEOPT, NBD_OPT_ABORT, 0)
             bytes_sent = client_socket.send(option_header)
             self.assertEqual(bytes_sent, 16)
 
@@ -110,23 +110,23 @@ class TestNBDServerNegotiation(unittest.TestCase):
             server_thread.start()
 
             client_socket.recv(18)
-            client_socket.sendall(struct.pack('>I', NBD_FLAG_C_FIXED_NEWSTYLE))
+            client_socket.sendall(struct.pack(">I", NBD_FLAG_C_FIXED_NEWSTYLE))
 
-            export_name = b'test'
-            option_data = struct.pack('>I', len(export_name)) + export_name + struct.pack('>H', 0)
-            option_header = struct.pack('>QII', IHAVEOPT, NBD_OPT_GO, len(option_data))
+            export_name = b"test"
+            option_data = struct.pack(">I", len(export_name)) + export_name + struct.pack(">H", 0)
+            option_header = struct.pack(">QII", IHAVEOPT, NBD_OPT_GO, len(option_data))
             client_socket.sendall(option_header + option_data)
 
             info_reply = client_socket.recv(32)
             self.assertEqual(len(info_reply), 32)
-            magic, option, reply_type, length = struct.unpack('>QIII', info_reply[:20])
+            magic, option, reply_type, length = struct.unpack(">QIII", info_reply[:20])
             self.assertEqual(magic, NBD_REP_MAGIC)
             self.assertEqual(option, NBD_OPT_GO)
             self.assertEqual(reply_type, NBD_REP_INFO)
 
             ack_reply = client_socket.recv(20)
             self.assertEqual(len(ack_reply), 20)
-            magic, option, reply_type, length = struct.unpack('>QIII', ack_reply)
+            magic, option, reply_type, length = struct.unpack(">QIII", ack_reply)
             self.assertEqual(magic, NBD_REP_MAGIC)
             self.assertEqual(option, NBD_OPT_GO)
             self.assertEqual(reply_type, NBD_REP_ACK)
@@ -154,14 +154,14 @@ class TestNBDServerNegotiation(unittest.TestCase):
             server_thread.start()
 
             client_socket.recv(18)
-            client_socket.sendall(struct.pack('>I', NBD_FLAG_C_FIXED_NEWSTYLE))
+            client_socket.sendall(struct.pack(">I", NBD_FLAG_C_FIXED_NEWSTYLE))
 
-            option_header = struct.pack('>QII', IHAVEOPT, NBD_OPT_ABORT, 0)
+            option_header = struct.pack(">QII", IHAVEOPT, NBD_OPT_ABORT, 0)
             client_socket.sendall(option_header)
 
             client_socket.settimeout(0.5)
             data = client_socket.recv(1024)
-            self.assertEqual(data, b'')
+            self.assertEqual(data, b"")
 
         finally:
             client_socket.close()
@@ -171,11 +171,11 @@ class TestNBDServerTransmission(unittest.TestCase):
 
     def _negotiate_and_enter_transmission(self, client_socket):
         client_socket.recv(18)
-        client_socket.sendall(struct.pack('>I', NBD_FLAG_C_FIXED_NEWSTYLE))
+        client_socket.sendall(struct.pack(">I", NBD_FLAG_C_FIXED_NEWSTYLE))
 
-        export_name = b'test'
-        option_data = struct.pack('>I', len(export_name)) + export_name + struct.pack('>H', 0)
-        option_header = struct.pack('>QII', IHAVEOPT, NBD_OPT_GO, len(option_data))
+        export_name = b"test"
+        option_data = struct.pack(">I", len(export_name)) + export_name + struct.pack(">H", 0)
+        option_header = struct.pack(">QII", IHAVEOPT, NBD_OPT_GO, len(option_data))
         client_socket.sendall(option_header + option_data)
 
         client_socket.recv(32)
@@ -183,7 +183,7 @@ class TestNBDServerTransmission(unittest.TestCase):
 
     def test_transmission_cmd_read(self):
         storage = InMemoryStorage()
-        test_data = b'hello world'
+        test_data = b"hello world"
         storage.write(0, test_data)
         server = NBDServer(storage)
 
@@ -206,11 +206,11 @@ class TestNBDServerTransmission(unittest.TestCase):
             handle = 12345
             offset = 0
             length = len(test_data)
-            cmd = struct.pack('>IHHQQL', NBD_REQUEST_MAGIC, 0, NBD_CMD_READ, handle, offset, length)
+            cmd = struct.pack(">IHHQQL", NBD_REQUEST_MAGIC, 0, NBD_CMD_READ, handle, offset, length)
             client_socket.sendall(cmd)
 
             reply = client_socket.recv(16)
-            magic, error, reply_handle = struct.unpack('>IIQ', reply)
+            magic, error, reply_handle = struct.unpack(">IIQ", reply)
             self.assertEqual(magic, NBD_SIMPLE_REPLY_MAGIC)
             self.assertEqual(error, 0)
             self.assertEqual(reply_handle, handle)
@@ -243,13 +243,15 @@ class TestNBDServerTransmission(unittest.TestCase):
 
             handle = 12346
             offset = 0
-            test_data = b'test write'
+            test_data = b"test write"
             length = len(test_data)
-            cmd = struct.pack('>IHHQQL', NBD_REQUEST_MAGIC, 0, NBD_CMD_WRITE, handle, offset, length)
+            cmd = struct.pack(
+                ">IHHQQL", NBD_REQUEST_MAGIC, 0, NBD_CMD_WRITE, handle, offset, length
+            )
             client_socket.sendall(cmd + test_data)
 
             reply = client_socket.recv(16)
-            magic, error, reply_handle = struct.unpack('>IIQ', reply)
+            magic, error, reply_handle = struct.unpack(">IIQ", reply)
             self.assertEqual(magic, NBD_SIMPLE_REPLY_MAGIC)
             self.assertEqual(error, 0)
             self.assertEqual(reply_handle, handle)
@@ -281,11 +283,11 @@ class TestNBDServerTransmission(unittest.TestCase):
             self._negotiate_and_enter_transmission(client_socket)
 
             handle = 12347
-            cmd = struct.pack('>IHHQQL', NBD_REQUEST_MAGIC, 0, NBD_CMD_FLUSH, handle, 0, 0)
+            cmd = struct.pack(">IHHQQL", NBD_REQUEST_MAGIC, 0, NBD_CMD_FLUSH, handle, 0, 0)
             client_socket.sendall(cmd)
 
             reply = client_socket.recv(16)
-            magic, error, reply_handle = struct.unpack('>IIQ', reply)
+            magic, error, reply_handle = struct.unpack(">IIQ", reply)
             self.assertEqual(magic, NBD_SIMPLE_REPLY_MAGIC)
             self.assertEqual(error, 0)
             self.assertEqual(reply_handle, handle)
@@ -314,16 +316,16 @@ class TestNBDServerTransmission(unittest.TestCase):
             self._negotiate_and_enter_transmission(client_socket)
 
             handle = 12348
-            cmd = struct.pack('>IHHQQL', NBD_REQUEST_MAGIC, 0, NBD_CMD_DISC, handle, 0, 0)
+            cmd = struct.pack(">IHHQQL", NBD_REQUEST_MAGIC, 0, NBD_CMD_DISC, handle, 0, 0)
             client_socket.sendall(cmd)
 
             client_socket.settimeout(0.5)
             data = client_socket.recv(1024)
-            self.assertEqual(data, b'')
+            self.assertEqual(data, b"")
 
         finally:
             client_socket.close()
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     unittest.main()
